fix: move wikibase/wikidata api stuff into core
